Issue #26: reorganize test modules for better decoupling
