+++ fencer.cabal
@@ -114,6 +114,8 @@ test-suite test-fencer
     Fencer.Logic.Test
     Fencer.Types.Test
     Fencer.Rules.Test
+    Fencer.Rules.Test.Helpers
+    Fencer.Rules.Test.Types
     Fencer.Server.Test
   default-language:
     Haskell2010
