+++ test/Fencer/Server/Test.hs
@@ -32,6 +32,8 @@ import           Fencer.Settings (defaultGRPCPort, getLogLevel, newLogger)
 import           Fencer.Types
 import           Fencer.Rules
 import qualified Fencer.Rules.Test as RTest
+import           Fencer.Rules.Test.Helpers (writeAndLoadRules)
+import           Fencer.Rules.Test.Types (RuleFile(..), simpleRuleFile)
 import qualified Fencer.Proto as Proto
 
 {-# ANN module ("HLint: ignore Reduce duplication" :: String) #-}
@@ -142,7 +144,7 @@ test_serverResponseReadPermissions =
     testCase "OK response with one YAML file without read permissions" $
       Temp.withSystemTempDirectory "fencer-config" $ \tempDir -> do
         server <- serverIO
-        RTest.writeAndLoadRules
+        writeAndLoadRules
           (#ignoreDotFiles False)
           (#root tempDir)
           (#files files)
@@ -158,13 +160,13 @@ test_serverResponseReadPermissions =
                 (expectedResponse, Grpc.StatusOk)
                 response
   where
-    files :: [RTest.RuleFile]
+    files :: [RuleFile]
     files =
-      [ RTest.MkRuleFile
+      [ MkRuleFile
           ("domain1" </> "config.yml")
           RTest.domain1Text
           (const Dir.emptyPermissions)
-      , RTest.simpleRuleFile
+      , simpleRuleFile
           ("domain2" </> "config" </> "config.yml")
           RTest.domain2Text
       ]
